builder/cloudstack: Do not process user data twice
The user_data field is currently processed twice. During the first
processing in Builder, all variables get expanded, including those set
which are only set later in stepDeployVirtualMachine, in particular the
address of the internal HTTP server. This change modifies the Builder to
only validate the user data so the processing is done only once in
stepDeployVirtualMachine when all variables are set.